Moved build system to modified Slashdev version
The build system involved here is described in fair detail at

http://www.slashdev.ca/2007/05/30/blackberry-development-with-ant-eclipse/

With some slight modifications to the common build to support building
with multiple JDE versions (specified as their three-digit version with
no dots).  This project will also work in that system with no
modification, as the only change visible at the project level is the
addition of the jde.supports property in project.properties.